+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [1.1.0] - 2021-12-27
+### Added
+- Alpine support.
+- Save with user timestamps information.
+- Supplemented the documentation.
+- Dispatch puller to everyone online user.
+
 ## [1.0.0] - 2021-12-25
 ### Added
 - License mit.